+++ packages/xy/src/index.tsx
@@ -1,2 +1,3 @@
+export * from './density'
 export * from './histogram'
 export * from './scatter'