+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Added
 
+-   [xy] components `Density`, `DensityCells`, `DensityCell`, and
+    `DensitySimpleCell` for creating density maps
 -   [xy] support for row-based and column-based datasets for `Scatter` views
 -   [core] function `cloneProps` to create shallow/deep copies of prop objects
 -   [core] wrapper `PersistentTooltipDataComponent` to create tooltips that